ce11_TSSs

Coordinates of promoter TSSs annotated in C. elegans (ce11) used in Serizay et al. 2020, "Tissue-specific profiling reveals distinctive regulatory architectures for ubiquitous, germline and somatic genes", BiorXiv.

data(ce11_TSSs)

Format

GRanges

Source

BiorXiv

References

Serizay et al. 2020, "Tissue-specific profiling reveals distinctive regulatory architectures for ubiquitous, germline and somatic genes", BiorXiv. (DOI)

Examples

data(ce11_TSSs)
lengths(ce11_TSSs)

#>    Ubiq. Germline  Neurons   Muscle   Hypod.  Intest. 
#>     2295     1098      889      984     1440     1512 

ce11_TSSs[[1]]

#> GRanges object with 2295 ranges and 0 metadata columns:
#>                                     seqnames            ranges strand
#>                                        <Rle>         <IRanges>  <Rle>
#>             rcor-1_chrI_26903_27053     chrI       26451-27350      -
#>           Y74C9A.1_chrI_43281_43431     chrI       42966-43865      +
#>           Y48G1C.6_chrI_86365_86515     chrI       85909-86808      -
#>   rpl-7,rab-11.1_chrI_110636_110786     chrI     110334-111233      +
#>   rpl-7,rab-11.1_chrI_110636_110786     chrI     110209-111108      -
#>                                 ...      ...               ...    ...
#>     zipt-7.2_chrX_15508115_15508265     chrX 15507666-15508565      -
#>       nipi-3_chrX_15606342_15606492     chrX 15606044-15606943      +
#>     F59D12.5_chrX_15647416_15647566     chrX 15646959-15647858      -
#>        cas-1_chrX_16814106_16814256     chrX 16813693-16814592      -
#>      C36E6.2_chrX_17451496_17451646     chrX 17451191-17452090      +
#>   -------
#>   seqinfo: 6 sequences from an unspecified genome; no seqlengths