Internal function

Internal function

chunkWrapper(chunk, genome, MOTIF, GENOME.WINDOW.SLIDING, BIN.WINDOW.SIZE,
  BIN.WINDOW.SLIDING, RANGE.FOR.SPECTRUM, FREQ)

Arguments

chunk
A GRanges from the output list obtained with getChunks

genome
A DNAStringSet object. Ideally, the sequence of an entire genome, obtained for instance by running Biostrings::getSeq( BSgenome.Celegans.UCSC.ce11::BSgenome.Celegans.UCSC.ce11 ).

MOTIF
String Oligonucleotide of interest.

GENOME.WINDOW.SLIDING
Integer The increment between bins over GRanges (default 2).

BIN.WINDOW.SIZE
Integer The width of the bins to split each primary bin in (default 60).

BIN.WINDOW.SLIDING
Integer The increment between secondary bins (default 5).

RANGE.FOR.SPECTRUM
Numeric vector The distances between nucleotides to take into consideration when performing Fast Fourier Transform (default seq_len(50)).

FREQ
Float The frequence of the dinucleotide to study (default 1/10).

Value

GRanges a GRanges object with a score column containing estimated periodicity