Function to generate a k-mer periodicity track

This function takes a set of GRanges in a genome, recover the corresponding sequences and divides them using a sliding window. For each sub-sequence, it then computes the PSD value of a k-mer of interest at a chosen period, and generates a linear .bigWig track from these values.

getPeriodicityTrack(
  genome = NULL,
  granges,
  motif = "WW",
  period = 10,
  cores = 12,
  extension = 1000,
  genome_sliding_size = 100,
  genome_sliding_sliding = 2,
  window_sliding_size = 100,
  window_sliding_bin = 5,
  range_spectrum = seq(5, 50),
  bw_file = NULL
)

Arguments

genome
DNAStringSet, BSgenome or genome ID

granges
GRanges object

motif
character, k-mer of interest.

period
Integer, the period of the dinucleotide to study (default=10).

cores
Integer, split the workload over several processors (default 12).

extension
Integer, the width the GRanges are going to be extended to (default 1000).

genome_sliding_size
Integer, the width of the bins to split the GRanges objects in (default 100).

genome_sliding_sliding
Integer, the increment between bins over GRanges (default 2).

window_sliding_size
Integer, the width of the bins to split each primary bin in (default 60).

window_sliding_bin
Integer, the increment between secondary bins (default 5).

range_spectrum
Numeric vector, the distances between nucleotides to take into consideration when performing Fast Fourier Transform (default seq_len(50)).

bw_file
character, the name of the output bigWig track

Value

Rlelist and a bigWig track in the working directory.

Examples

data(ce11_proms)
getPeriodicityTrack(
    genome = 'ce11',
    ce11_proms[1],
    motif = 'WW',
    period = 10,
    cores = 1
)

#> 
#>     The genome has been divided in 100-bp long windows 
#>     with a sliding window of 2-bp.
#>     
#>     549 windows overlap the 1 input loci (extended to 1000-bp).
#>     
#>     The mapping will be split among 1 jobs. Each job will 
#>     analyse 549 windows.
#>     
#>     Now starting...
#>     

#> >> 1 bins computed /// TIME: 2020-06-12 15:59:52

#> >> 549 bins computed /// TIME: 2020-06-12 16:00:32

#> 
#>    SUCCESS: WW-periodicity_g-100^2_b-100^5.bw has been created!

#>    1098 bases covered by the generated track.

#> RleList of length 7
#> $chrI
#> numeric-Rle of length 15072434 with 412 runs
#>   Lengths:    10798        2        4        2 ...        2        2 15060538
#>   Values :   0.0000   9.1421   7.6238   6.8686 ...   8.3510  12.0618   0.0000
#> 
#> $chrII
#> numeric-Rle of length 15279421 with 1 run
#>   Lengths: 15279421
#>   Values :        0
#> 
#> $chrIII
#> numeric-Rle of length 13783801 with 1 run
#>   Lengths: 13783801
#>   Values :        0
#> 
#> $chrIV
#> numeric-Rle of length 17493829 with 1 run
#>   Lengths: 17493829
#>   Values :        0
#> 
#> $chrV
#> numeric-Rle of length 20924180 with 1 run
#>   Lengths: 20924180
#>   Values :        0
#> 
#> ...
#> <2 more elements>